requirements.txt
# Core Framework
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

--extra-index-url https://download.pytorch.org/whl/cpu

# Web Framework
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.20
aiofiles==25.1.0

# AI/ML Models
timm==1.0.20
facenet-pytorch==2.6.0
transformers==4.57.1
huggingface-hub==0.35.3
accelerate==1.10.1
safetensors==0.6.2

# Computer Vision
opencv-python==4.10.0.84
opencv-contrib-python==4.9.0.80
opencv-python-headless==4.9.0.80
face-recognition==1.3.0
face-recognition-models==0.3.0
mtcnn==1.0.0
dlib==20.0.0

# Image Processing
Pillow==10.2.0
imageio==2.37.0
albumentations==2.0.8
albucore==0.0.24

# Video Processing
moviepy==2.2.1
imageio-ffmpeg==0.6.0
librosa==0.11.0
soundfile==0.13.1

# Data Science
numpy==1.26.4
pandas==2.3.3
scipy==1.15.3
scikit-learn==1.7.2

# Visualization
matplotlib==3.10.7
seaborn==0.13.2

# Utilities
tqdm==4.67.1
python-dotenv==1.1.1
requests==2.32.5
certifi==2025.10.5
pydantic==2.12.2